requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Django==3.1.6
django-filter==21.1
djangorestframework==3.12.4
idna==3.2
lxml==4.6.3
Markdown==3.3.4
multitasking==0.0.9
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
pycparser==2.20
PyJWT==2.2.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
yfinance==0.1.59